+++ src/check/analyze-import-access/analyze.ts
