+++ src/check/analyze-import-access/analyzed-result.ts
