+++ src/check/analyze-import-access/index.ts
