+++ src/get-all-ts-files/index.ts
@@ -1,0 +1,1 @@
+export { getAllTsFiles } from "./get-all-ts-files";
