+++ src/import-config/index.ts
@@ -1,0 +1,1 @@
+export { importConfig } from "./import-config";
